+++ src/server/importer/rpc.test.ts
@@ -156,8 +156,10 @@ describe('RPC: Import Functions', () => {
       const result = importCSV(N26ImportMock, 'N26');
 
       expect(importDataSpy).toHaveBeenCalled();
-      expect((result as { success: true; message: string }).message).toBe('imported 4 rows!');
       expect(result.success).toBe(true);
+      if (result.success) {
+        expect(result.message).toBe('imported 4 rows!');
+      }
     });
 
     test('is able to handle rabobank import', () => {
@@ -169,8 +171,10 @@ describe('RPC: Import Functions', () => {
       const result = importCSV(raboImportMock, 'rabobank');
 
       expect(importDataSpy).toHaveBeenCalled();
-      expect((result as { success: true; message: string }).message).toBe('imported 1 rows!');
       expect(result.success).toBe(true);
+      if (result.success) {
+        expect(result.message).toBe('imported 1 rows!');
+      }
     });
 
     test('is able to handle bank of america', () => {
@@ -194,8 +198,10 @@ describe('RPC: Import Functions', () => {
       expect(importDataSpy).toHaveBeenCalledWith(expect.arrayContaining([
         expect.arrayContaining([new Date('2023-09-12'),-100,'Utility Bill Payment']),
       ]), undefined);
-      expect((result as { success: true; message: string }).message).toBe('imported 5 rows!')
       expect(result.success).toBe(true);
+      if (result.success) {
+        expect(result.message).toBe('imported 5 rows!')
+      }
     })
 
     test('sorts by date before importing', () => {